+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Updated the feed card to remove the time stamp.
 - Updated the home feed to no longer have the first card be a hero card. It's now
 a regular card just like everything else.
+### Added
+- Added a "Prompt Modal" to prompt users of necessary information.
+- Push notifications!
+- Support for dynamic links
 ### Fixed
 - Fixed an issue with the createDataChannel not existing which caused the app
 to break on the latest Microsoft Edge releases.
 
-## [5.1.0] - UNRELEASED
+## [5.1.0]
 ### Fixed
 - fixed cc exp date validation
 - Unpredictable custom tag in give/now with schedules
@@ -23,6 +27,7 @@ to break on the latest Microsoft Edge releases.
 ### Added
 - Added back the ability to save payments without a transaction.
 - Platform and version headers to api requests
+
 
 ## [5.0.22]
 ### Fixed